refactor: 패키지 구조 변경에 따른 리팩토링
