+++ src/main/java/com/prgrms/amabnb/reservation/dto/request/CreateReservationRequest.java
@@ -2,7 +2,7 @@ package com.prgrms.amabnb.reservation.dto.request;
 
 import java.time.LocalDate;
 
-import com.prgrms.amabnb.common.model.Money;
+import com.prgrms.amabnb.common.vo.Money;
 import com.prgrms.amabnb.reservation.entity.Reservation;
 import com.prgrms.amabnb.reservation.entity.vo.ReservationDate;
 import com.prgrms.amabnb.room.entity.Room;
