+++ src/test/java/com/prgrms/amabnb/reservation/repository/ReservationRepositoryTest.java
@@ -14,7 +14,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 
 import com.prgrms.amabnb.common.RepositoryTest;
-import com.prgrms.amabnb.common.model.Money;
+import com.prgrms.amabnb.common.vo.Email;
+import com.prgrms.amabnb.common.vo.Money;
+import com.prgrms.amabnb.common.vo.PhoneNumber;
 import com.prgrms.amabnb.reservation.entity.Reservation;
 import com.prgrms.amabnb.reservation.entity.vo.ReservationDate;
 import com.prgrms.amabnb.room.entity.Room;
@@ -25,8 +27,6 @@ import com.prgrms.amabnb.room.entity.vo.RoomOption;
 import com.prgrms.amabnb.room.repository.CreateRoomRepository;
 import com.prgrms.amabnb.user.entity.User;
 import com.prgrms.amabnb.user.entity.UserRole;
-import com.prgrms.amabnb.user.entity.vo.Email;
-import com.prgrms.amabnb.user.entity.vo.PhoneNumber;
 import com.prgrms.amabnb.user.repository.UserRepository;
 
 class ReservationRepositoryTest extends RepositoryTest {
