+++ src/test/java/com/prgrms/amabnb/room/repository/RoomRepositoryTest.java
@@ -7,9 +7,8 @@ import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.context.annotation.Import;
 
+import com.prgrms.amabnb.common.RepositoryTest;
 import com.prgrms.amabnb.common.vo.Money;
 import com.prgrms.amabnb.room.entity.Room;
 import com.prgrms.amabnb.room.entity.RoomScope;
@@ -17,9 +16,7 @@ import com.prgrms.amabnb.room.entity.RoomType;
 import com.prgrms.amabnb.room.entity.vo.RoomAddress;
 import com.prgrms.amabnb.room.entity.vo.RoomOption;
 
-@DataJpaTest
-@Import(RoomTestConfig.class)
-class RoomRepositoryTest {
+class CreateRoomRepositoryTest extends RepositoryTest {
 
     private final RoomAddress roomAddress = new RoomAddress("00000", "창원", "의창구");
     private final Money price = new Money(20000);
